Struct libunbound::Context

pub struct Context { /* private fields */ }

Implementations

impl Context

The validation context is created to hold the resolver status, validation keys and a small cache (containing messages, rrsets, roundtrip times, trusted keys, lameness information).

pub fn new() -> Result<Self, Error>

Create a resolving and validation context. The information from /etc/resolv.conf and /etc/hosts is not utilised by default. Use ub_ctx_resolvconf and ub_ctx_hosts to read them.

pub fn resolve( &self, name: &str, rrtype: RecordType, rrclass: DNSClass ) -> Result<Answer, Error>

Perform resolution and validation of the target name. The context is finalized, and can no longer accept config changes. @param name: domain name in text format (a zero terminated text string). @param rrtype: type of RR @param rrclass: class of RR

pub fn set_option(&self, opt: &str, value: &str) -> Result<(), Error>

Set an option for the context @param opt: option name from the unbound.conf config file format. (not all settings applicable). The name includes the trailing ‘:’ for example

let ctx = libunbound::Context::new().unwrap();
ctx.set_option("logfile:", "mylog.txt");

This is a power-users interface that lets you specify all sorts of options.

For some specific options, such as adding trust anchors, special routines exist.

@param val: value of the option.

pub fn get_option(&self, opt: &str) -> Result<String, Error>

Get an option from the context. @param opt: option name from the unbound.conf config file format. (not all settings applicable). The name excludes the trailing ‘:’ for example:

let ctx = libunbound::Context::new().unwrap();
ctx.get_option("logfile");

This is a power-users interface that lets you specify all sorts of options. In cases with multiple entries (auto-trust-anchor-file), a newline delimited list is returned in the string.

pub fn load_unbound_config_file( &self, config_file_name: &str ) -> Result<(), Error>

setup configuration for the given context. @param config_file_name: unbound config file (not all settings applicable). This is a power-users interface that lets you specify all sorts of options. For some specific options, such as adding trust anchors, special routines exist.

pub fn set_forward(&self, addr: Option<IpAddr>) -> Result<(), Error>

Set machine to forward DNS queries to, the caching resolver to use. IP4 or IP6 address. Forwards all DNS requests to that machine, which is expected to run a recursive resolver. If the proxy is not DNSSEC-capable, validation may fail. Can be called several times, in that case the addresses are used as backup servers. To read the list of nameservers from /etc/resolv.conf (from DHCP or so), use the call ub_ctx_resolvconf. At this time it is only possible to set configuration before the\n\tfirst resolve is done. If the addr is None, forwarding is disabled.

pub fn set_tls(&self, tls: bool) -> Result<(), Error>

Use DNS over TLS to send queries to machines set with set_forward(). At this time it is only possible to set configuration before the first resolve is done. @param tls: enable or disable DNS over TLS

pub fn set_stub( &self, zone: &str, addr: Option<IpAddr>, is_prime: bool ) -> Result<(), Error>

Add a stub zone, with given address to send to. This is for custom root hints or pointing to a local authoritative dns server. For dns resolvers and the ‘DHCP DNS’ ip address, use set_forward. This is similar to a stub-zone entry in unbound.conf.

It is only possible to set configuration before the first resolve is done. @param zone: name of the zone, string. @param addr: address, IP4 or IP6 in string format. The addr is added to the list of stub-addresses if the entry exists. If the addr is None the stub entry is removed. @param isprime: set to true to set stub-prime to yes for the stub. For local authoritative servers, people usually set it to false, For root hints it should be set to true.

pub fn load_resolv_conf(&self, file_name: Option<&str>) -> Result<(), Error>

Read list of nameservers to use from the filename given. Usually "/etc/resolv.conf". Uses those nameservers as caching proxies. If they do not support DNSSEC, validation may fail. Only nameservers are picked up, the searchdomain, ndots and other settings from resolv.conf(5) are ignored. At this time it is only possible to set configuration before the first resolve is done. n @param fname: file name string. If None "/etc/resolv.conf" is used.

pub fn load_hosts(&self, file_name: Option<&str>) -> Result<(), Error>

Read list of hosts from the filename given. Usually "/etc/hosts" These addresses are not flagged as DNSSEC secure when queried for. At this time it is only possible to set configuration before the first resolve is done. @param fname: file name string. If None "/etc/hosts" is used.

pub fn add_trust_anchor(&self, ta: &str) -> Result<(), Error>

Add a trust anchor to the given context. The trust anchor is a string, on one line, that holds a valid DNSKEY or DS RR. tAt this time it is only possible to add trusted keys before the first resolve is done.

@param ta: string, with zone-format RR on one line.

[domainname] [TTL optional] [type] [class optional] [rdata contents]

pub fn add_builtin_trust_anchors(&self) -> Result<(), Error>

Add the built-in trust anchors to the context. These anchors were correct at the time that this module was authored, however, for robust DNSSEC, you should consider bootstrapping an anchor file and loading it via load_trust_anchor_file() if you want stronger assertions that these anchors are correct and to keep them updated.

pub fn load_trust_anchor_file(&self, file_name: &str) -> Result<(), Error>

Add trust anchors to the given context. Pass name of a file with DS and DNSKEY records (like from dig or drill). At this time it is only possible to add trusted keys before the first resolve is done. @param fname: filename of file with keyfile with trust anchors.

pub fn bootstrap_trust_anchor_file(&self, file_name: &str) -> Result<(), Error>

Use file_name to maintain root trust anchors for DNSSEC. If the file doesn’t already exist, it will be populated with the built-in ROOT_TRUST_ANCHORS defined by this module.

pub fn load_trust_anchor_file_with_auto_update( &self, file_name: &str ) -> Result<(), Error>

Add trust anchor to the given context that is tracked with RFC5011 automated trust anchor maintenance. The file is written to when the trust anchor is changed. Pass the name of a file that was output from eg. unbound-anchor, or you can start it by providing a trusted DNSKEY or DS record on one line in the file. At this time it is only possible to add trusted keys before the first resolve is done. @param fname: filename of file with trust anchor.

pub fn set_debug_output(&self, stream: *mut FILE) -> Result<(), Error>

Set debug output (and error output) to the specified stream. Pass NULL to disable. Default is stderr. @param out: FILE* out file stream to log to.

pub fn set_debug_level(&self, level: DebugLevel) -> Result<(), Error>

Set debug verbosity for the context Output is directed to stderr or whatever was configured via set_debug_output(). @param d: debug level, 0 is off, 1 is very minimal, 2 is detailed, and 3 is lots.

pub fn print_local_zones(&self) -> Result<(), Error>

Debug routine. Print the local zone information to debug output. Is finalized by the routine.

pub fn zone_add(&self, zone_name: &str, zone_type: &str) -> Result<(), Error>

Add a new zone with the zonetype to the local authority info of the library. Is finalized by the routine. @param zone_name: name of the zone in text, "example.com" tIf it already exists, the type is updated. @param zone_type: type of the zone (like for unbound.conf) in text.

pub fn zone_remove(&self, zone_name: &str) -> Result<(), Error>

Remove zone from local authority info of the library. Is finalized by the routine. @param zone_name: name of the zone in text, "example.com\" If it does not exist, nothing happens.

pub fn add_local_data(&self, data: &str) -> Result<(), Error>

Add localdata to the library local authority info. Similar to local-data config statement. @param data: the resource record in text format, for example "www.example.com IN A 127.0.0.1"

pub fn remove_local_data(&self, data: &str) -> Result<(), Error>

Remove localdata from the library local authority info. @param data: the name to delete all data from, like `“www.example.com”

pub fn into_async(self) -> Result<AsyncContext, Error>

Convert the context into an async resolving context. A helper thread will be spawned to manage the context, and will automatically shut down when the returned AsyncContext is dropped.